requirements.txt
google-generativeai==0.7.1
pillow==10.3.0
python-dotenv==1.0.1